+++ api/deplist_test.go
@@ -1,26 +1,26 @@
-package deplist_test
+package api_test
 
 import (
+	"github.com/elgohr/deplist/api"
 	"os"
 	"path/filepath"
 	"reflect"
 	"testing"
-	"github.com/elgohr/deplist/deplist"
 )
 
 var testCases = []struct {
 	name   string
 	dir    string
-	o      deplist.Opts
+	o      api.Opts
 	output []string
 }{
 	{"c", "/", 0, nil},
 	{"b", "/", 0, []string{"c"}},
 	{"a", "/", 0, []string{"b", "c"}},
 	{".", "testdata/src/a", 0, []string{"b", "c"}},
-	{"a", "/", deplist.OptTestImports, []string{"b", "c", "d"}},
-	{"a", "/", deplist.OptStd, []string{"b", "c", "unsafe"}},
-	{"a", "/", deplist.OptTestImports | deplist.OptStd, []string{"b", "c", "d", "unsafe"}},
+	{"a", "/", api.OptTestImports, []string{"b", "c", "d"}},
+	{"a", "/", api.OptStd, []string{"b", "c", "unsafe"}},
+	{"a", "/", api.OptTestImports | api.OptStd, []string{"b", "c", "d", "unsafe"}},
 	{"e", "/", 0, []string{"e/vendor/v0", "e/vendor/v0/vendor/a"}},
 }
 
@@ -30,7 +30,7 @@ func TestFindDeps(t *testing.T) {
 		t.Fatal(err)
 	}
 	for _, tt := range testCases {
-		deps, err := deplist.FindDeps(tt.name, tt.dir, filepath.Join(cwd, "testdata"), tt.o)
+		deps, err := api.FindDeps(tt.name, tt.dir, filepath.Join(cwd, "testdata"), tt.o)
 		if err != nil {
 			t.Fatal(err)
 		}
